+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements
 
+ * `global_parameters` is a new command group for operations on [global runtime parameters](https://www.rabbitmq.com/docs/parameters)
  * `nodes` is a new command group for operations on nodes
  * `parameters` is a new command group for operations on [runtime parameters](https://www.rabbitmq.com/docs/parameters)
  * `users` is a new command group for operations on users